web-app/requirements.txt
python-dotenv
Flask
pinecone-client 
openai 
tiktoken 
langchain